requirements.txt
streamlit==1.28.0
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
sqlite3==2.6.0
python-dotenv==1.0.0
requests==2.31.0
openai==0.28.1
faker==18.13.0
surprise==1.1.3